fix: Обновлены conftest и тесты для совместимости с новой архитектурой - Исправлен conftest.py для поддержки старого и нового API - Фикс dtype в mock embedder (float32 вместо float64) - Все тесты Фазы 1 проходят (11/11) ✅ - Старые тесты не сломаны (используют fallback на старый API)

+++ tests/test_phase_1_architecture.py
@@ -235,10 +235,10 @@ class TestSOLIDPrinciples:
         # Мок-реализации
         class FakeEmbedder(BaseEmbedder):
             def embed_documents(self, texts):
-                return [np.zeros(768) for _ in texts]
+                return [np.zeros(768, dtype=np.float32) for _ in texts]
             
             def embed_query(self, text):
-                return np.zeros(768)
+                return np.zeros(768, dtype=np.float32)
         
         class FakeSplitter(BaseSplitter):
             def split(self, document):
